+++ lib/byebug/interfaces/local_interface.rb
@@ -15,13 +15,13 @@ module Byebug
     end
 
     #
-    # Reads a single line of input using Readline. If Ctrl-D is pressed, it
+    # Reads a single line of input using Reline. If Ctrl-D is pressed, it
     # returns "continue", meaning that program's execution will go on.
     #
     # @param prompt Prompt to be displayed.
     #
     def readline(prompt)
-      with_repl_like_sigint { Readline.readline(prompt) || EOF_ALIAS }
+      with_repl_like_sigint { Reline.readline(prompt) || EOF_ALIAS }
     end
 
     #