+++ test/minitest_runner_test.rb
@@ -56,8 +56,10 @@ module Byebug
     private
 
     def run_minitest_runner(*args)
+      env = shell_out_env(simplecov: false)
+
       Bundler.with_original_env do
-        output, status = Open3.capture2e(shell_out_env(simplecov: false), *binstub, *args)
+        output, status = Open3.capture2e(env, *binstub, *args)
 
         assert_equal true, status.success?, output